src/main/java/com/core/spring/domain/order/OrderServiceImpl.java
package com.core.spring.domain.order;

import com.core.spring.domain.member.Member;
import com.core.spring.domain.member.MemberRepository;
import com.core.spring.domain.member.MemoryMemberRepository;

public class OrderServiceImpl implements OrderService{
    private MemberRepository memberRepository = new MemoryMemberRepository();
    private DiscountPolicy discountPolicy = new RateDiscountPolicy();

    @Override
    public Order createOrder(Long memberId, String itemName, int itemPrice) {
        Member member = memberRepository.findById(memberId);
        int discount = discountPolicy.discount(member, itemPrice);

        return new Order(memberId,itemName,itemPrice,discount);
    }
}
